Assignment 8: Binary Search Trees initial commit
The Problem

We have been asked to construct a program that spell-checks a document
given a "dictionary" of valid words.

We know that the key to an efficient spell-checker is an efficient
mechanism for storing and searching the dictionary. We decide to
maintain the dictionary (simply a list of strings) in sorted order, so
that searching the list will be efficient. Furthermore, we decide to
store the dictionary in a binary search tree (BST) because all
operations on the BST are potentially very fast.

We must first implement a BST of generically-typed items. The
implementation should include a class to represent a binary tree node
and a class to represent a BST (a collection of binary tree nodes).
Please note that there is no requirement that the BST be balanced.

We must then apply the BST to the problem of spell-checking a document,
using a BST to store a dictionary of strings and looking up each
document string in the BST. The implementation should include a class
to represent a spell-checker and a class to demonstrate the
spell-checker for various dictionary and document files.